# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Messaging service API documentation",
        "version": "1"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Current version"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authenticate user and get bearer token to use with API.",
                "description": "Creates new access token and invalidates previously issued tokens if any",
                "operationId": "72bf786ad8ff53c7391309c836350a8d",
                "requestBody": {
                    "description": "User's credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User's email address",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Access Token",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Invalidate all tokens issued to user.",
                "description": "Invalidate all tokens issued to user",
                "operationId": "5660f1ae61b95f0283dfc205d5ffdad5",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}